+++ test/lib/parse/query/constraints/exists_test.rb
@@ -17,4 +17,25 @@ class TestExistsConstraint < Minitest::Test
     {"field" => { @key => value } }
   end
 
+  def test_scalar_values
+
+  [true, false].each do |value|
+    constraint = @klass.new(:field, value)
+    expected = build(value).as_json
+    assert_equal expected, constraint.build.as_json
+  end
+
+  ["true", 1, nil].each do |value|
+    constraint = @klass.new(:field, value)
+
+    assert_raises(ParseConstraintError) do
+      expected = build(value).as_json
+      constraint.build.as_json
+    end
+
+
+  end
+
+end
+
 end