+++ test/lib/parse/query/constraints/nullability_test.rb
@@ -20,5 +20,21 @@ class TestNullabilityConstraint < Minitest::Test
     end
   end
 
+  def test_scalar_values
+    [true, false].each do |value|
+      constraint = @klass.new(:field, value)
+      expected = build(value).as_json
+      assert_equal expected, constraint.build.as_json
+    end
+
+    ["true", 1, nil].each do |value|
+      constraint = @klass.new(:field, value)
+      assert_raises(ParseConstraintError) do
+        expected = build(value).as_json
+        constraint.build.as_json
+      end
+    end
+
+  end
 
 end